code review changes - move ref_ct outside of rwlock, break up/refactor types mod, misc cleanup
