+++ src/lib.rs
@@ -1,7 +1,6 @@
 #![deny(warnings)]
 mod telemetry;
 mod telemetry_subscriber;
-mod types;
 mod visitor;
 
 #[cfg(test)]
@@ -10,5 +9,5 @@ mod visitor;
 extern crate lazy_static;
 
 pub use crate::telemetry::HoneycombTelemetry;
+pub use crate::telemetry::TraceId;
 pub use crate::telemetry_subscriber::TelemetrySubscriber;
-pub use crate::types::TraceId;
