+++ src/telemetry.rs
@@ -1,5 +1,5 @@
-use crate::types::{Event, Span};
-use libhoney::FieldHolder;
+use chrono::{DateTime, Utc};
+use libhoney::{json, FieldHolder};
 use std::collections::HashMap;
 use std::sync::Mutex;
 
@@ -49,10 +49,207 @@ impl Telemetry for HoneycombTelemetry {
     }
 }
 
+#[derive(PartialEq, Eq, Hash, Clone, Debug)]
+pub struct SpanId(tracing::Id, u64);
+
+impl std::fmt::Display for SpanId {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}-{}", self.0.into_u64(), self.1)
+    }
+}
+
+impl SpanId {
+    pub(crate) fn from_id(id: tracing::Id, instance_id: u64) -> Self {
+        SpanId(id, instance_id)
+    }
+}
+
+#[derive(PartialEq, Eq, Hash, Clone, Debug)]
+pub struct TraceId(String);
+
+impl TraceId {
+    pub fn record_on_current_span(self) {
+        let mut id = Some(self);
+        tracing::dispatcher::get_default(|d| {
+            if let Some(s) = d.downcast_ref::<crate::telemetry_subscriber::TelemetrySubscriber>() {
+                // required b/c get_default takes FnMut, however we know it will only be invoked once
+                let id = id.take().expect("fn should not be invoked twice");
+                s.record_trace_id(id)
+            } else {
+                // TODO: does this merit a panic? probably yes.
+                panic!("unable to record TraceId, this thread does not have TelemetrySubscriber registered as default")
+            }
+        })
+    }
+
+    pub fn new(u: String) -> Self {
+        TraceId(u)
+    }
+    pub(crate) fn generate() -> Self {
+        use rand::Rng;
+
+        let u: u128 = rand::thread_rng().gen();
+        TraceId(format!("trace-{}", u))
+    }
+}
+
+#[derive(Debug, Clone)]
+pub(crate) struct Span<'a> {
+    pub(crate) id: SpanId,
+    pub(crate) trace_id: TraceId,
+    pub(crate) parent_id: Option<SpanId>,
+    pub(crate) initialized_at: DateTime<Utc>,
+    pub(crate) elapsed_ms: i64,
+    pub(crate) level: tracing::Level,
+    pub(crate) name: &'a str,
+    pub(crate) target: &'a str,
+    pub(crate) service_name: &'a str,
+    pub(crate) values: HashMap<String, libhoney::Value>, // bag of misc values
+}
+
+impl<'a> Span<'a> {
+    #[cfg(test)]
+    pub(crate) fn into_static(self) -> Span<'static> {
+        let e: Span<'static> = Span {
+            name: test::lift_to_static(self.name),
+            target: test::lift_to_static(self.target),
+            service_name: test::lift_to_static(self.service_name),
+            id: self.id,
+            trace_id: self.trace_id,
+            parent_id: self.parent_id,
+            initialized_at: self.initialized_at,
+            elapsed_ms: self.elapsed_ms,
+            level: self.level,
+            values: self.values,
+        };
+        e
+    }
+
+    pub(crate) fn into_values(self) -> HashMap<String, libhoney::Value> {
+        let mut values = self.values;
+
+        values.insert(
+            // magic honeycomb string (trace.span_id)
+            "trace.span_id".to_string(),
+            json!(format!("span-{}", self.id.to_string())),
+        );
+
+        values.insert(
+            // magic honeycomb string (trace.trace_id)
+            "trace.trace_id".to_string(),
+            // using explicit trace id passed in from ctx (req'd for lazy eval)
+            json!(self.trace_id.0),
+        );
+
+        values.insert(
+            // magic honeycomb string (trace.parent_id)
+            "trace.parent_id".to_string(),
+            self.parent_id
+                .map(|pid| json!(format!("span-{}", pid.to_string())))
+                .unwrap_or(json!(null)),
+        );
+
+        // magic honeycomb string (service_name)
+        values.insert("service_name".to_string(), json!(self.service_name));
+
+        values.insert("level".to_string(), json!(format!("{}", self.level)));
+
+        values.insert(
+            "timestamp".to_string(),
+            json!(self.initialized_at.to_rfc3339()),
+        );
+
+        // not honeycomb-special but tracing-provided
+        values.insert("name".to_string(), json!(self.name));
+        values.insert("target".to_string(), json!(self.target));
+
+        // honeycomb-special (I think, todo: get full list of known values)
+        values.insert("duration_ms".to_string(), json!(self.elapsed_ms));
+
+        values
+    }
+}
+
+// copy strings into lazy static ref in tests
+
+#[derive(Clone, Debug)]
+pub(crate) struct Event<'a> {
+    pub(crate) trace_id: TraceId,
+    pub(crate) parent_id: Option<SpanId>,
+    pub(crate) initialized_at: DateTime<Utc>,
+    pub(crate) level: tracing::Level,
+    pub(crate) name: &'a str,
+    pub(crate) target: &'a str,
+    pub(crate) service_name: &'a str,
+    pub(crate) values: HashMap<String, libhoney::Value>, // bag of misc values
+}
+
+impl<'a> Event<'a> {
+    #[cfg(test)]
+    pub(crate) fn into_static(self) -> Event<'static> {
+        let e: Event<'static> = Event {
+            name: test::lift_to_static(self.name),
+            target: test::lift_to_static(self.target),
+            service_name: test::lift_to_static(self.service_name),
+            trace_id: self.trace_id,
+            parent_id: self.parent_id,
+            initialized_at: self.initialized_at,
+            level: self.level,
+            values: self.values,
+        };
+        e
+    }
+
+    pub(crate) fn into_values(self) -> HashMap<String, libhoney::Value> {
+        let mut values = self.values;
+
+        values.insert(
+            // magic honeycomb string (trace.trace_id)
+            "trace.trace_id".to_string(),
+            // using explicit trace id passed in from ctx (req'd for lazy eval)
+            json!(self.trace_id.0),
+        );
+
+        values.insert(
+            // magic honeycomb string (trace.parent_id)
+            "trace.parent_id".to_string(),
+            self.parent_id
+                .map(|pid| json!(format!("span-{}", pid.to_string())))
+                .unwrap_or(json!(null)),
+        );
+
+        // magic honeycomb string (service_name)
+        values.insert("service_name".to_string(), json!(self.service_name));
+
+        values.insert("level".to_string(), json!(format!("{}", self.level)));
+
+        values.insert(
+            "timestamp".to_string(),
+            json!(self.initialized_at.to_rfc3339()),
+        );
+
+        // not honeycomb-special but tracing-provided
+        values.insert("name".to_string(), json!(self.name));
+        values.insert("target".to_string(), json!(self.target));
+
+        values
+    }
+}
+
 #[cfg(test)]
 pub(crate) mod test {
     use super::*;
     use std::sync::Arc;
+
+    pub(super) fn lift_to_static(s: &'_ str) -> &'static str {
+        use aovec::Aovec;
+        lazy_static! {
+            static ref STATIC_STRING_STORAGE: Aovec<String> = Aovec::new(256);
+        }
+
+        let idx = STATIC_STRING_STORAGE.push(s.to_string());
+        STATIC_STRING_STORAGE.get(idx).unwrap()
+    }
 
     /// Mock telemetry capability
     pub(crate) struct TestTelemetry {
